Add `forwardUserName` functionality to ANCM V2.
When using IIS to proxy access to apps running in Docker containers (or other non-native apps),
the app needs to know the identify of the authenticated IIS user. With Windows authentication
enabled in IIS, ANCM V2 can be configured to forward the user's *token*, which unfortunately
an app running in a container cannot use.

This commit enables passing the authenticated Windows user *name* in a request header
(configurable, defaults to `MS-ASPNETCORE-USER`). By default, the full authenticated user name
is passed (e.g. `DOMAIN\User`), but the option `forwardUserDomain="false"` excludes the
domain, passing only `User` instead.

Ref https://msdn.microsoft.com/en-us/library/ms689371(v=vs.90).aspx